build: move sonar polyfill to shared lib
