+++ apps/vue-app/.nvmrc
@@ -1,1 +1,1 @@
-20
+22
