+++ apps/vue-app/.storybook/main.js
@@ -11,7 +11,6 @@ const config = {
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-docs',
-    '@chromatic-com/storybook',
     '@storybook/addon-a11y',
     {
       name: '@storybook/addon-coverage',
