+++ apps/vue-app/documentation/dashboard-widgets.png
