+++ apps/vue-app/documentation/open-architecture.png
