+++ apps/vue-app/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "engines": {
-    "node": ">=18",
+    "node": ">=22",
     "pnpm": ">=9"
   },
   "scripts": {
@@ -24,25 +24,23 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "lodash": "^4.17.21",
-    "vue": "^3.5.17",
-    "vue3-toastify": "^0.2.8"
+    "@creedengo/sonar-services": "workspace:*",
+    "vue": "^3.5.17"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^4.0.1",
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@creedengo/eslint-plugin": "^2.1.0",
     "@eslint/compat": "^1.1.1",
     "@eslint/js": "^9.30.1",
     "@playwright/test": "^1.53.1",
-    "@storybook/addon-a11y": "9.0.15",
+    "@storybook/addon-a11y": "^9.0.15",
     "@storybook/addon-coverage": "^1.0.4",
     "@storybook/addon-docs": "^9.0.15",
-    "@storybook/addon-links": "9.0.15",
+    "@storybook/addon-links": "^9.0.15",
     "@storybook/builder-vite": "^9.0.15",
-    "@storybook/vue3": "9.0.15",
-    "@storybook/vue3-vite": "9.0.15",
+    "@storybook/vue3": "^9.0.15",
+    "@storybook/vue3-vite": "^9.0.15",
     "@vitejs/plugin-vue": "^5.2.4",
     "@vitest/coverage-v8": "^3.2.4",
     "@vue/test-utils": "^2.4.6",
@@ -58,9 +56,9 @@
     "jsdom": "^24.1.0",
     "msw": "^2.10.2",
     "msw-storybook-addon": "^2.0.5",
-    "prettier": "3.5.3",
+    "prettier": "^3.5.3",
     "sonarqube-scanner": "^4.3.0",
-    "storybook": "9.0.15",
+    "storybook": "^9.0.15",
     "vite": "^6.3.5",
     "vite-plugin-vue-devtools": "^7.7.7",
     "vitest": "^3.2.4"
