+++ apps/vue-app/src/api/sonar/sonar-i18n.js
@@ -1,5 +1,5 @@
 
-import SonarI18nPolyfill from './polyfills/sonar-i18n'
+import { SonarI18nPolyfill } from '@creedengo/sonar-services'
 
 export const t = self.t ?? SonarI18nPolyfill.t
 export const tp = self.tp ?? SonarI18nPolyfill.tp
