+++ shared/sonar-services/src/index.js
@@ -1,0 +1,4 @@
+export { default } from './polyfills/sonar-request'
+export * from './polyfills/sonar-request'
+export * from './polyfills/sonar-i18n'
+export * from './polyfills/sonar-toast'
