+++ shared/sonar-services/src/polyfills/sonar-i18n.spec.js
