+++ shared/sonar-services/src/polyfills/sonar-request.spec.js
