+++ shared/sonar-services/vitest.config.js
@@ -1,0 +1,18 @@
+import { fileURLToPath } from 'node:url'
+import { mergeConfig, defineConfig, configDefaults } from 'vitest/config'
+import viteConfig from './vite.config'
+
+export default mergeConfig(
+  viteConfig,
+  defineConfig({
+    test: {
+      //environment: 'jsdom',
+      exclude: [...configDefaults.exclude],
+      root: fileURLToPath(new URL('./', import.meta.url)),
+      coverage: {
+        reporter: ['text', 'lcov'],
+        exclude: ['.*rc*','*.config.js'],
+      }
+    }
+  })
+)
